+++ src/main/webapp/headerEmployee.jsp
@@ -389,7 +389,7 @@
                                 <li class="flex">
                                     <a
                                             class="inline-flex items-center w-full px-2 py-1 text-sm font-semibold transition-colors duration-150 rounded-md hover:bg-gray-100 hover:text-gray-800 dark:hover:bg-gray-800 dark:hover:text-gray-200"
-                                            href="/manager-page"
+                                            href="/auths"
                                     >
                                         <svg
                                                 class="w-4 h-4 mr-3"
