+++ src/main/webapp/jsp/menuAdmin/add-task.jsp
@@ -1,0 +1,12 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: hcmqu
+  Date: 1/4/2024
+  Time: 10:06 AM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page pageEncoding="UTF-8" contentType="text/html;charset=UTF-8" language="java" %>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<c:import url="../../headerAdmin.jsp" />
+    <h3>Add - Task</h3>
+<c:import url="../../footerAdmin.jsp" />
