+++ src/main/webapp/jsp/pagesIndex/indexEmployee.jsp
@@ -6,11 +6,9 @@
   Time: 8:24 PM
   To change this template use File | Settings | File Templates.
 --%>
-<c:import url="../../header.jsp" />
 
+<c:import url="../../headerEmployee.jsp"/>
+    <h3>${member.fullName}</h3>
+    <input type="hidden" id="fullName" value="${member.fullName}">
+<c:import url="../../footerAdmin.jsp"/>
 
-<h3>${member.fullName}</h3>
-<input type="hidden" id="fullName" value="${member.fullName}">
-
-
-<c:import url="../../footer.jsp" />
